@@ -22,16 +22,15 @@ android {
         }
     }
 
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
 
     defaultConfig {
         applicationId "com.mononz.crypz"
         minSdkVersion 21
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode (System.getenv('CIRCLE_BUILD_NUM') != null ? Integer.valueOf(System.getenv('CIRCLE_BUILD_NUM')) : 1)
         versionName "1.0.0"
-        testInstrumentationRunner 'android.support.test.runner.AndroidJUnitRunner'
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         vectorDrawables.useSupportLibrary = true
 
         javaCompileOptions {
