@@ -1,9 +1,9 @@
 package com.mononz.crypz.data.local
 
 import android.app.Application
-import android.arch.persistence.room.Database
-import android.arch.persistence.room.Room
-import android.arch.persistence.room.RoomDatabase
+import androidx.room.Database
+import androidx.room.Room
+import androidx.room.RoomDatabase
 import com.mononz.crypz.data.local.dao.*
 
 import com.mononz.crypz.data.local.entity.*
